run formatter

+++ Jax/Routes/LogReg.php
@@ -71,10 +71,7 @@ final class LogReg implements Route
             'logout' => $this->logout(),
             'toggleInvisible' => $this->toggleInvisible(),
             'forgotPassword' => $this->forgotPassword(),
-            default => $this->login(
-                $this->request->asString->post('user'),
-                $this->request->asString->post('pass'),
-            ),
+            default => $this->login($this->request->asString->post('user'), $this->request->asString->post('pass')),
         };
     }
 
@@ -146,21 +143,12 @@ final class LogReg implements Route
                 . ' administrator.',
             !$name || !$dispname => 'Name and display name required.',
             $pass1 !== $pass2 => 'The passwords do not match.',
-            mb_strlen($dispname) > 30 || mb_strlen(
-                $name,
-            ) > 30 => 'Display name and username must be under 30 characters.',
-            ($badNameChars && preg_match($badNameChars, $name))
-                || $this->textFormatting->blockhtml(
-                    $name,
-                ) !== $name => 'Invalid characters in username!',
-            $badNameChars && preg_match(
-                $badNameChars,
-                $dispname,
-            ) => 'Invalid characters in display name!',
-            !filter_var(
-                $email,
-                FILTER_VALIDATE_EMAIL,
-            ) => "That isn't a valid email!",
+            mb_strlen($dispname) > 30 || mb_strlen($name) > 30
+                => 'Display name and username must be under 30 characters.',
+            $badNameChars && preg_match($badNameChars, $name) || $this->textFormatting->blockhtml($name) !== $name
+                => 'Invalid characters in username!',
+            $badNameChars && preg_match($badNameChars, $dispname) => 'Invalid characters in display name!',
+            !filter_var($email, FILTER_VALIDATE_EMAIL) => "That isn't a valid email!",
             $this->ipAddress->isBanned() => 'You have been banned from registering on this board.',
             !$this->didPassCaptcha() => 'You did not pass the captcha. Are you a bot?',
             default => null,
@@ -168,20 +156,13 @@ final class LogReg implements Route
 
         if ($error !== null) {
             $this->page->command('error', $error);
-            $this->page->append(
-                'PAGE',
-                $this->template->render('error', ['message' => $error]),
-            );
+            $this->page->append('PAGE', $this->template->render('error', ['message' => $error]));
 
             return;
         }
 
         // Are they attempting to use an existing username/display name?
-        $member = Member::selectOne(
-            'WHERE `name`=? OR `displayName`=?',
-            $name,
-            $dispname,
-        );
+        $member = Member::selectOne('WHERE `name`=? OR `displayName`=?', $name, $dispname);
 
         $error = match (true) {
             $member?->name === $name => 'That username is taken!',
@@ -191,10 +172,7 @@ final class LogReg implements Route
 
         if ($error !== null) {
             $this->page->command('error', $error);
-            $this->page->append(
-                'PAGE',
-                $this->template->render('error', ['message' => $error]),
-            );
+            $this->page->append('PAGE', $this->template->render('error', ['message' => $error]));
 
             return;
         }
@@ -202,17 +180,12 @@ final class LogReg implements Route
         $newMember = new Member();
         $newMember->displayName = $dispname;
         $newMember->email = $email;
-        $newMember->groupID = $this->config->getSetting('membervalidation')
-            ? 5
-            : 1;
+        $newMember->groupID = $this->config->getSetting('membervalidation') ? 5 : 1;
         $newMember->ip = $this->ipAddress->asBinary() ?? '';
         $newMember->joinDate = $this->database->datetime();
         $newMember->lastVisit = $this->database->datetime();
         $newMember->name = $name;
-        $newMember->pass = password_hash(
-            $pass1,
-            PASSWORD_DEFAULT,
-        );
+        $newMember->pass = password_hash($pass1, PASSWORD_DEFAULT);
         $newMember->insert();
 
         $stats = Stats::selectOne();
@@ -225,19 +198,14 @@ final class LogReg implements Route
         $this->login($name, $pass1);
     }
 
-    private function login(
-        ?string $username = null,
-        #[\SensitiveParameter] ?string $password = null,
-    ): void {
+    private function login(?string $username = null, #[\SensitiveParameter] ?string $password = null): void
+    {
         if ($username && $password) {
             if ($this->session->get()->isBot !== 0) {
                 return;
             }
 
-            $member = Member::selectOne(
-                'WHERE `name`=?',
-                $username,
-            );
+            $member = Member::selectOne('WHERE `name`=?', $username);
 
             $this->user->login($member->id ?? null, $password);
 
@@ -246,26 +214,17 @@ final class LogReg implements Route
                     $this->page->command('closewindow', '#loginform');
                 }
 
-                $this->session->setPHPSessionValue(
-                    'uid',
-                    $this->user->get()->id,
-                );
+                $this->session->setPHPSessionValue('uid', $this->user->get()->id);
                 $loginToken = base64_encode(openssl_random_pseudo_bytes(128));
 
                 $token = new Token();
-                $token->expires = $this->database->datetime(
-                    Carbon::now('UTC')->addMonth()->getTimestamp(),
-                );
+                $token->expires = $this->database->datetime(Carbon::now('UTC')->addMonth()->getTimestamp());
                 $token->token = $loginToken;
                 $token->type = 'login';
                 $token->uid = $this->user->get()->id;
                 $token->insert();
 
-                $this->request->setCookie(
-                    'utoken',
-                    $loginToken,
-                    Carbon::now('UTC')->addMonth()->getTimestamp(),
-                );
+                $this->request->setCookie('utoken', $loginToken, Carbon::now('UTC')->addMonth()->getTimestamp());
                 $this->session->clean($this->user->get()->id);
                 $this->session->set('uid', $this->user->get()->id);
                 if ($this->registering) {
@@ -278,23 +237,16 @@ final class LogReg implements Route
                     return;
                 }
             } else {
-                $this->page->append(
-                    'PAGE',
-                    $this->template->render(
-                        'error',
-                        ['message' => 'Incorrect username/password'],
-                    ),
-                );
+                $this->page->append('PAGE', $this->template->render('error', [
+                    'message' => 'Incorrect username/password',
+                ]));
                 $this->page->command('error', 'Incorrect username/password');
             }
 
             $this->session->erase('location');
         }
 
-        $this->page->append(
-            'PAGE',
-            $this->template->render('logreg/login-form'),
-        );
+        $this->page->append('PAGE', $this->template->render('logreg/login-form'));
     }
 
     private function logout(): void
@@ -303,11 +255,7 @@ final class LogReg implements Route
         // to maintain users online.
         $uToken = $this->request->cookie('utoken');
         if ($uToken !== null) {
-            $this->database->delete(
-                'tokens',
-                'WHERE `token`=?',
-                $uToken,
-            );
+            $this->database->delete('tokens', 'WHERE `token`=?', $uToken);
             $this->request->setCookie('utoken', null, -1);
         }
 
@@ -315,27 +263,10 @@ final class LogReg implements Route
         $this->session->applyChanges();
         session_unset();
         session_destroy();
-        $this->template->reset(
-            'USERBOX',
-            $this->template->render(
-                'global/userbox-logged-out',
-            ),
-        );
-        $this->page->command(
-            'update',
-            'userbox',
-            $this->template->render(
-                'global/userbox-logged-out',
-            ),
-        );
+        $this->template->reset('USERBOX', $this->template->render('global/userbox-logged-out'));
+        $this->page->command('update', 'userbox', $this->template->render('global/userbox-logged-out'));
         $this->page->command('preventNavigation');
-        $this->page->append(
-            'PAGE',
-            $this->template->render(
-                'success',
-                ['message' => 'Logged out successfully'],
-            ),
-        );
+        $this->page->append('PAGE', $this->template->render('success', ['message' => 'Logged out successfully']));
         if ($this->request->isJSAccess()) {
             return;
         }
@@ -349,19 +280,13 @@ final class LogReg implements Route
 
         $this->session->applyChanges();
 
-        $this->page->command(
-            'setstatus',
-            $this->session->get()->hide !== 0 ? 'invisible' : 'online',
-        );
+        $this->page->command('setstatus', $this->session->get()->hide !== 0 ? 'invisible' : 'online');
         $this->page->command('preventNavigation');
     }
 
     private function forgotPasswordHasToken(string $tokenId): ?string
     {
-        $token = Token::selectOne(
-            'WHERE `token`=? AND expires>=NOW()',
-            $tokenId,
-        );
+        $token = Token::selectOne('WHERE `token`=? AND expires>=NOW()', $tokenId);
 
         $pass1 = $this->request->asString->post('pass1');
         $pass2 = $this->request->asString->post('pass2');
@@ -385,18 +310,11 @@ final class LogReg implements Route
             return 'The associated account could not be found';
         }
 
-        $member->pass = password_hash(
-            $pass1,
-            PASSWORD_DEFAULT,
-        );
+        $member->pass = password_hash($pass1, PASSWORD_DEFAULT);
         $member->update();
 
         // Delete all forgotpassword tokens for this user.
-        $this->database->delete(
-            'tokens',
-            "WHERE `uid`=? AND `type`='forgotpassword'",
-            $token->uid,
-        );
+        $this->database->delete('tokens', "WHERE `uid`=? AND `type`='forgotpassword'", $token->uid);
 
         // Just making use of the way
         // registration redirects to the index.
@@ -422,90 +340,67 @@ final class LogReg implements Route
 
         if ($tokenId !== null && $tokenId !== '') {
             $error = $this->forgotPasswordHasToken($tokenId);
-            $page .= ($error !== null ? $this->page->error($error) : '')
-                . $this->template->render(
-                    'logreg/forgot-password2-form',
-                    [
-                        'hiddenFields' => Template::hiddenFormFields(
-                            [
-                                'id' => $tokenId,
-                                'uid' => $uid ?? '',
-                            ],
-                        ),
-                    ],
-                );
+            $page .=
+                ($error !== null ? $this->page->error($error) : '')
+                . $this->template->render('logreg/forgot-password2-form', [
+                    'hiddenFields' => Template::hiddenFormFields([
+                        'id' => $tokenId,
+                        'uid' => $uid ?? '',
+                    ]),
+                ]);
         } else {
             if ($user) {
-                $member = Member::selectOne(
-                    'WHERE `name`=?',
-                    $user,
-                );
+                $member = Member::selectOne('WHERE `name`=?', $user);
 
                 if ($member === null) {
                     $error = "There is no user registered as <strong>{$user}</strong>, sure this is correct?";
-                    $page .= $this->template->render(
-                        'error',
-                        ['message' => $error],
-                    );
+                    $page .= $this->template->render('error', ['message' => $error]);
                 } else {
                     // Generate token.
-                    $forgotpasswordtoken
-                        = base64_encode(openssl_random_pseudo_bytes(128));
+                    $forgotpasswordtoken = base64_encode(openssl_random_pseudo_bytes(128));
 
                     $token = new Token();
-                    $token->expires = $this->database->datetime(
-                        Carbon::now('UTC')->getTimestamp() + 3600 * 24,
-                    );
+                    $token->expires = $this->database->datetime(Carbon::now('UTC')->getTimestamp() + (3600 * 24));
                     $token->token = $forgotpasswordtoken;
                     $token->type = 'forgotpassword';
                     $token->uid = $member->id;
                     $token->insert();
 
-                    $link = $this->router->getRootURL() . $this->router->url(
-                        'forgotPassword',
-                        [
+                    $link =
+                        $this->router->getRootURL()
+                        . $this->router->url('forgotPassword', [
                             'uid' => $member->id,
                             'tokenId' => rawurlencode($forgotpasswordtoken),
-                        ],
-                    );
-                    $mailResult = $this->mailer->mail(
-                        $member->email,
-                        'Recover Your Password!',
-                        <<<HTML
-                            You have received this email because a password
-                            request was received at {BOARDLINK}
-                            <br>
-                            <br>
-                            If you did not request a password change, simply
-                            ignore this email and no actions will be taken.
-                            If you would like to change your password, please
-                            visit the following page and follow the on-screen
-                            instructions:
-                            <a href='{$link}'>{$link}</a>
-                            <br>
-                            <br>
-                            Thanks!
-                            HTML,
-                    );
+                        ]);
+                    $mailResult = $this->mailer->mail($member->email, 'Recover Your Password!', <<<HTML
+                        You have received this email because a password
+                        request was received at {BOARDLINK}
+                        <br>
+                        <br>
+                        If you did not request a password change, simply
+                        ignore this email and no actions will be taken.
+                        If you would like to change your password, please
+                        visit the following page and follow the on-screen
+                        instructions:
+                        <a href='{$link}'>{$link}</a>
+                        <br>
+                        <br>
+                        Thanks!
+                        HTML);
 
                     if (!$mailResult) {
-                        $page .= $this->template->render(
-                            'error',
-                            [
-                                'message' => 'There was a problem sending the email. '
-                                    . 'Please contact the administrator.',
-                            ],
-                        );
+                        $page .= $this->template->render('error', [
+                            'message' =>
+                                'There was a problem sending the email. ' . 'Please contact the administrator.',
+                        ]);
                     } else {
-                        $page .= $this->template->render(
-                            'success',
-                            [
-                                'message' => 'An email has been sent to the email associated '
+                        $page .= $this->template->render('success', [
+                            'message' =>
+                                'An email has been sent to the email associated '
                                     . 'with this account. Please check your email and '
                                     . 'follow the instructions in order to recover '
                                     . 'your password.',
-                            ],
-                        );
+                        ]);
                     }
                 }
             }
